+++ src/User/User.php
@@ -19,6 +19,8 @@ final class User
     private $firstname;
     /** @var string */
     private $lastname;
+    /** @var \DateTimeImmutable|null */
+    private $birthday;
     /** @var UserAddress|null */
     private $billingAddress;
     /** @var UserAddress|null */
@@ -34,6 +36,7 @@ final class User
         $this->title = empty($data['title']) ? null : new UserTitle($data['title']);
         $this->firstname = (string) $data['firstName'];
         $this->lastname = (string) $data['lastName'];
+        $this->birthday = empty($data['birthday']) ? null : new \DateTimeImmutable($data['birthday']);
         $this->billingAddress = isset($data['addresses']['billing']) ? new UserAddress($data['addresses']['billing']) : null;
         $this->shippingAddress = isset($data['addresses']['shipping']) ? new UserAddress($data['addresses']['shipping']) : null;
     }
@@ -61,6 +64,11 @@ final class User
     public function getLastname(): string
     {
         return $this->lastname;
+    }
+
+    public function getBirthday(): ?\DateTimeImmutable
+    {
+        return $this->birthday;
     }
 
     public function getBillingAddress(): ?UserAddress