+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - Added `\Wizaplace\Catalog\Declination::isBrandNew`
 - Added `\Wizaplace\Catalog\Product::getVideo`
+- Added `\Wizaplace\SDK\User\User::getBirthday`
+- Added `\Wizaplace\SDK\User\UpdateUserCommand::setBirthday`
 
 ### Bugfixes